Fix SetWindowLong issue

+++ Flow.Launcher/MainWindow.xaml.cs
@@ -377,13 +377,13 @@ namespace Flow.Launcher
 
         private IntPtr WndProc(IntPtr hwnd, int msg, IntPtr wParam, IntPtr lParam, ref bool handled)
         {
-            if (Win32Helper.WM_ENTERSIZEMOVE(msg))
+            if (Win32Helper.WM_ENTERSIZEMOVE((uint)msg))
             {
                 _initialWidth = (int)Width;
                 _initialHeight = (int)Height;
                 handled = true;
             }
-            else if (Win32Helper.WM_EXITSIZEMOVE(msg))
+            else if (Win32Helper.WM_EXITSIZEMOVE((uint)msg))
             {
                 if (_initialHeight != (int)Height)
                 {
